src/test/java/com/amazon/randomcutforest/sampler/WeightedPointTest.java
/*
 * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License").
 * You may not use this file except in compliance with the License.
 * A copy of the License is located at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * or in the "license" file accompanying this file. This file is distributed
 * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
 * express or implied. See the License for the specific language governing
 * permissions and limitations under the License.
 */

package com.amazon.randomcutforest.sampler;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class WeightedPointTest {

	@Test
	public void testNew() {
		WeightedPoint point = new WeightedPoint(new double[]{0.99, -55.2}, 999L, 1.23);
		assertEquals(1.23, point.getWeight());
		assertEquals(999L, point.getSequenceIndex());
		assertArrayEquals(new double[]{0.99, -55.2}, point.getPoint());
	}

}
